@@ -2,18 +2,13 @@
 
 namespace STS\ZipStream\Events;
 
-use STS\ZipStream\ZipStream;
+use STS\ZipStream\Builder;
+use ZipStream\ZipStream;
 
 class ZipStreamed
 {
-    /** @var ZipStream */
-    public $zip;
-
-    /**
-     * @param ZipStream $zip
-     */
-    public function __construct(ZipStream $zip)
+    public function __construct(public Builder $builder, public ZipStream $zip, public int $size)
     {
-        $this->zip = $zip;
+
     }
 }
