@@ -3,11 +3,10 @@
 namespace STS\ZipStream\Tests;
 
 use Illuminate\Support\Str;
-use STS\ZipStream\ZipStream;
-use Zip;
+use STS\ZipStream\Builder;
 use ZipArchive;
 use Orchestra\Testbench\TestCase;
-use STS\ZipStream\ZipStreamFacade;
+use STS\ZipStream\Facades\Zip;
 use STS\ZipStream\ZipStreamServiceProvider;
 
 class ZipTest extends TestCase
@@ -20,7 +19,7 @@ class ZipTest extends TestCase
     protected function getPackageAliases($app)
     {
         return [
-            'Zip' => ZipStreamFacade::class
+            'Zip' => Zip::class
         ];
     }
 
@@ -30,17 +29,14 @@ class ZipTest extends TestCase
         file_put_contents("/tmp/test1.txt", "this is the first test file for test run $testrun");
         file_put_contents("/tmp/test2.txt", "this is the second test file for test run $testrun");
 
-        /** @var ZipStream $zip */
+        /** @var Builder $zip */
         $zip = Zip::create("small.zip", ["/tmp/test1.txt", "/tmp/test2.txt"]);
-        $sizePrediction = $zip->predictZipSize();
 
         // Create a random folder path that doesn't exist, so we know it was created
         $dir = "/tmp/" . Str::random();
         $zip->saveTo($dir);
 
-        $this->assertFalse($zip->opt->isEnableZip64());
         $this->assertTrue(file_exists("$dir/small.zip"));
-        $this->assertEquals($sizePrediction, filesize("$dir/small.zip"));
 
 		$z = new ZipArchive();
         $z->open("$dir/small.zip");
@@ -60,14 +56,11 @@ class ZipTest extends TestCase
         exec('dd if=/dev/zero count=5120 bs=1048576 >/tmp/bigfile.txt');
         exec('dd if=/dev/zero count=1024 bs=1048576 >/tmp/medfile.txt');
 
-        /** @var ZipStream $zip */
+        /** @var Builder $zip */
         $zip = Zip::create("large.zip", ["/tmp/test1.txt", "/tmp/test2.txt", "/tmp/bigfile.txt", "/tmp/medfile.txt"]);
-        $sizePrediction = $zip->predictZipSize();
         $zip->saveTo("/tmp");
 
-        $this->assertTrue($zip->opt->isEnableZip64());
         $this->assertTrue(file_exists("/tmp/large.zip"));
-        $this->assertEquals($sizePrediction, filesize("/tmp/large.zip"));
 
         $z = new ZipArchive();
         $z->open("/tmp/large.zip");
